Remove AccessPathExpression in favor of AccessPathIdentifier.
The two classes are virtually identical and I don't see any conceptual difference
between the two, at least not in the current state of the heap analysis.

An AccessPathExpression does not contain any nested expressions.
It seems perfectly reasonable for the FieldAccess semantics to yield
an AccessPathIdentifier, analogous to Variable yielding a VariableIdentifier.