Use List[Option[String]] as access paths in EdgeLocalIdentifiers for consistency.
This means that EdgeLocalIdentifier is now in sync with Edge,
which also uses None to indicate the field of a local variable edge.
So we need to treat local variable edges separately on fewer occasions.
In fact, Path.condition now also recurses over the local variable edge.

That is, access paths are now empty if and only if the EdgeLocalIdentifier
refers to a field of the source vertex.