Make whole-program analysis great again
The forward interproc interpreter now expectes the whole program. All methods are (in top-down) enqueued to the workinglinst. Methods are then analyzed the following way:
- method is never called -> start analyzing with arguments = top
- method is called by other methods -> analyze with arguments = calling context. if there's no calling context (yet) arguments = bottom

Possible issues: Need to check if we'll reach the widening limit faster. A method may be enqueued a couple of times until all entryStates (calling contexts) are available.