Bugfix for bottom-up analysis
When analysis a program in which some methods form a connected component, then it's necessary to reanalayse the callers of the same connected component again. See the last example in Evaluation - Specification Inference of the project report. Without reanalyinng the caller again here the bottom-up inference would come up with a too strong postcondition that cannot be proven by a verifier.